correção progresso e ajustes finais

+++ server/routes/api.php
@@ -44,22 +44,22 @@ Route::post('profile-photo','UserController@changeProfilePhoto')->middleware('au
 //update name and email of USER
 Route::put('user-update', 'UserController@updateUser')->middleware('auth:api');
 //top five
-Route::get('top-users', 'UserController@getTopFive');
+Route::get('top-users', 'UserController@getTopFive')->middleware('auth:api');
 
 Route::get('categories', 'CategoryController@index')
     ->middleware('auth:api');
-Route::post('categories', 'CategoryController@store');
-Route::get('categories/{id}', 'CategoryController@getById');
-Route::delete('categories/{id}', 'CategoryController@destroy');
-Route::put('categories/{id}', 'CategoryController@update');
+Route::post('categories', 'CategoryController@store')->middleware('auth:api','adminRole');
+Route::get('categories/{id}', 'CategoryController@getById')->middleware('auth:api');
+Route::delete('categories/{id}', 'CategoryController@destroy')->middleware('auth:api','adminRole');
+Route::put('categories/{id}', 'CategoryController@update')->middleware('auth:api','adminRole');
 
 Route::get('themes', 'ThemeController@index')
-    ->middleware('auth:api','adminRole');
-Route::get('themes/{id}', 'ThemeController@getById');
-Route::get('themes-by-category/{id}', 'ThemeController@themeByCategory');
-Route::post('themes', 'ThemeController@store');
-Route::delete('themes/{id}', 'ThemeController@destroy');
-Route::put('themes/{id}', 'ThemeController@update');
+    ->middleware('auth:api');
+Route::get('themes/{id}', 'ThemeController@getById')->middleware('auth:api');
+Route::get('themes-by-category/{id}', 'ThemeController@themeByCategory')->middleware('auth:api');
+Route::post('themes', 'ThemeController@store')->middleware('auth:api','adminRole')->middleware('auth:api','adminRole');
+Route::delete('themes/{id}', 'ThemeController@destroy')->middleware('auth:api','adminRole');
+Route::put('themes/{id}', 'ThemeController@update')->middleware('auth:api','adminRole');
 
 
 Route::post('teste', 'TestController@file');
@@ -67,15 +67,15 @@ Route::delete('teste', 'TestController@destroy');
 
 //content-solicitation
 Route::get('content-solicitations', 'ContentSolicitationController@index')->middleware('auth:api');
-Route::get('content-solicitations/{id}', 'ContentSolicitationController@getById');
+Route::get('content-solicitations/{id}', 'ContentSolicitationController@getById')->middleware('auth:api');
 Route::post('content-solicitations', 'ContentSolicitationController@store')->middleware('auth:api');
-Route::post('content-solicitations/teste', 'ContentSolicitationController@update');
-Route::delete('content-solicitations/{id}', 'ContentSolicitationController@destroy');
+Route::post('content-solicitations/teste', 'ContentSolicitationController@update')->middleware('auth:api');
+Route::delete('content-solicitations/{id}', 'ContentSolicitationController@destroy')->middleware('auth:api');
 Route::get('content-by-user', 'ContentSolicitationController@getContentsByUser')->middleware('auth:api');
 //Content status control
-Route::post('content-change-status/{id}', 'ContentSolicitationController@contentChangeStatus');
+Route::post('content-change-status/{id}', 'ContentSolicitationController@contentChangeStatus')->middleware('auth:api');
 //last contents
-Route::get('lastest-contents', 'ContentSolicitationController@getLastestContents');
+Route::get('lastest-contents', 'ContentSolicitationController@getLastestContents')->middleware('auth:api');
 
 //download files
 Route::get('/downloadFile/{id}','FileController@download');
@@ -86,20 +86,20 @@ Route::get('/downloadVideo/{id}', 'VideoController@download');
 Route::get('courses', 'CourseController@index')->middleware('auth:api');
 Route::post('courses', 'CourseController@store')->middleware('auth:api');
 Route::get('courses/{id}', 'CourseController@getById')->middleware('auth:api');
-Route::put('courses/{id}', 'CourseController@update');
-Route::delete('courses/{id}', 'CourseController@destroy');
-Route::post('send-to-approve/{id}', 'CourseController@sendToApprove');
-Route::post('change-status/{id}', 'CourseController@changeStatus');
+Route::put('courses/{id}', 'CourseController@update')->middleware('auth:api');
+Route::delete('courses/{id}', 'CourseController@destroy')->middleware('auth:api');
+Route::post('send-to-approve/{id}', 'CourseController@sendToApprove')->middleware('auth:api');
+Route::post('change-status/{id}', 'CourseController@changeStatus')->middleware('auth:api');
 Route::get('courses-by-user', 'CourseController@allCoursesByUser')->middleware('auth:api');
-Route::get('course-full-contents/{id}', 'CourseController@getCourseFullContents');
+Route::get('course-full-contents/{id}', 'CourseController@getCourseFullContents')->middleware('auth:api');
 //last courses
-Route::get('lastest-courses','CourseController@getLastestCourses');
+Route::get('lastest-courses','CourseController@getLastestCourses')->middleware('auth:api');
 
 //content to courses
 Route::post('add-content-in-course', 'ContentToCourseController@addContent')->middleware('auth:api');
-Route::get('contents-by-course/{id}', 'ContentToCourseController@contentsByCourse');
-Route::delete('contents-by-course/{id}', 'ContentToCourseController@destroy');
-Route::post('change-position', 'ContentToCourseController@changePosition');
+Route::get('contents-by-course/{id}', 'ContentToCourseController@contentsByCourse')->middleware('auth:api');
+Route::delete('contents-by-course/{id}', 'ContentToCourseController@destroy')->middleware('auth:api');
+Route::post('change-position', 'ContentToCourseController@changePosition')->middleware('auth:api');
 //With status done or not
 Route::get('contents-by-course-is-done/{id}', 'ContentToCourseController@contentsByCourseWithProgress')
     ->middleware('auth:api');
